requirements.txt
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
pdfplumber==0.10.3
requests==2.31.0
python-docx==1.0.1 